requirements.txt
requests
mega.py
